requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
#asyncio==3.4.3
attrs==24.2.0
azure-functions==1.21.1
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
#logging==0.4.9.6
multidict==6.1.0
propcache==0.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
urllib3==2.2.3
yarl==1.14.0
